build: move sqs-to-converter's cleanup fn to external +
- move sqs-to-converter's cleanup fn to external and add simplest test
- fix typo on docs

Ref. metriport/metriport-internal#1698
