+++ packages/lambdas/src/sqs-to-converter.ts
@@ -10,6 +10,7 @@ import { Log, prefixedLog } from "./shared/log";
 import { apiClient } from "./shared/oss-api";
 import { S3Utils } from "./shared/s3";
 import { SQSUtils } from "./shared/sqs";
+import { cleanUpPayload } from "./sqs-to-converter/cleanup";
 
 // Keep this as early on the file as possible
 capture.init();
@@ -390,21 +391,4 @@ async function sendConversionResult(
     },
   };
   await sqsUtils.sqs.sendMessage(sendParams).promise();
-}
-function cleanUpPayload(payloadRaw: string): string {
-  const payloadNoCDUNK = removeCDUNK(payloadRaw);
-  const payloadNoNullFlavor = removeNullFlavor(payloadNoCDUNK);
-  return payloadNoNullFlavor;
-}
-
-function removeCDUNK(payloadRaw: string): string {
-  const stringToReplace = /xsi:type="CD UNK"/g;
-  const replacement = `xsi:type="CD"`;
-  return payloadRaw.replace(stringToReplace, replacement);
-}
-
-function removeNullFlavor(payloadRaw: string): string {
-  const stringToReplace = /<id\s*nullFlavor\s*=\s*".*?"\s*\/>/g;
-  const replacement = `<id extension="1" root="1"/>`;
-  return payloadRaw.replace(stringToReplace, replacement);
 }
